Directory restructure, and made application a simple singleton
Making it singleton allows the base Application class to have a static get app function to get a polymorphic pointer to the real app, and it only really makes sense to have one app anyway, for the most part.  This also includes a directory restructure that allows us to use platform/windows/header.hh instead of ../platform etc.  it's a minor directory structure change