Changed clear after window update
window update effectively created a new frame, and clear is intended to be the first action in that frame.  When this is reversed, then clear is the last action of the previous frame, and all you will ever see is the clear color.  this was confusing and hard to figure out the solution to